feat(middleware-sdk-s3): restructure folder
