+++ packages/client-s3-browser/commands/UploadPartCopyCommand.ts
@@ -1,6 +1,6 @@
 import * as __aws_sdk_bucket_endpoint_middleware from "@aws-sdk/bucket-endpoint-middleware";
 import * as __aws_sdk_md5_js from "@aws-sdk/md5-js";
-import * as __aws_sdk_middleware_s3_copysource from "@aws-sdk/middleware-sdk-s3";
+import * as __aws_sdk_middleware_sdk_s3 from "@aws-sdk/middleware-sdk-s3";
 import * as __aws_sdk_middleware_stack from "@aws-sdk/middleware-stack";
 import * as __aws_sdk_ssec_middleware from "@aws-sdk/ssec-middleware";
 import * as __aws_sdk_types from "@aws-sdk/types";
@@ -81,7 +81,7 @@ export class UploadPartCopyCommand
         priority: 0
       }
     );
-    stack.add(__aws_sdk_middleware_s3_copysource.encodeCopySource, {
+    stack.add(__aws_sdk_middleware_sdk_s3.encodeCopySource, {
       step: "initialize",
       priority: 0,
       tags: { COPYSOURCE: true, URI_ENCODE: true }
