+++ packages/middleware-sdk-s3/src/encode-copy-source.spec.ts
